Move `ty::{Term, TermKind}` to their own little module (cute)
